refactor(mobile): split swap tests into files
