--- apps/mobile/src/features/swap/swap-state/use-swap-state.spec.ts
@@ -1,2209 +1,0 @@
-import { determineSwapExecutability } from '@/features/swap/swap-state/hooks/use-swap-executability';
-import { renderHookWithProviders } from '@/tests/test-utils';
-import { STACKS_MAINNET } from '@stacks/network';
-import { act, waitFor } from '@testing-library/react';
-import { assert, describe, expect, it, vi } from 'vitest';
-
-import { MarketData, SwapQuote } from '@leather.io/models';
-import { AccountSwapAsset } from '@leather.io/services';
-import { createMoney } from '@leather.io/utils';
-
-import {
-  createAccountRequest,
-  createAccountSwapAsset,
-  createSwapQuote,
-  defaultBtcAsset,
-  defaultSbtcAsset,
-  defaultStxAsset,
-} from './test-utils/fixtures';
-import {
-  createStubBitcoinPayer,
-  createStubBitcoinTransactionFeesService,
-  createStubMarketDataService,
-  createStubNetwork,
-  createStubStacksSigner,
-  createStubStacksTransactionFeesService,
-  createStubSwapService,
-} from './test-utils/services.stub';
-import { UseSwapStateProps, useSwapState } from './use-swap-state';
-
-vi.mock('@/hooks/use-debounced-value', () => ({
-  useDebouncedValue: <T>(value: T) => value,
-}));
-
-interface RenderUseSwapStateParams extends Omit<UseSwapStateProps, 'dependencies'> {
-  baseSwapAssets?: AccountSwapAsset[];
-  targetSwapAssets?: AccountSwapAsset[];
-  swapQuotes?: SwapQuote[];
-  marketData: MarketData;
-}
-
-function renderUseSwapState({
-  quoteCurrencyPreference = 'USD',
-  baseSwapAssets,
-  targetSwapAssets,
-  swapQuotes,
-  marketData,
-  ...rest
-}: Partial<RenderUseSwapStateParams> = {}) {
-  const { result } = renderHookWithProviders(() =>
-    useSwapState({
-      quoteCurrencyPreference,
-      dependencies: {
-        accountRequest: createAccountRequest(),
-        services: {
-          swapService: createStubSwapService({ baseSwapAssets, targetSwapAssets, swapQuotes }),
-          marketDataService: createStubMarketDataService({ marketData }),
-          bitcoinTransactionFeesService: createStubBitcoinTransactionFeesService(),
-          stacksTransactionFeesService: createStubStacksTransactionFeesService(),
-        },
-        stacks: {
-          stacksSigner: createStubStacksSigner(),
-          stacksNetwork: STACKS_MAINNET,
-          broadcast: () => Promise.resolve({ ok: true, txid: 'test-txid' }) as any,
-          nextNonce: undefined,
-        },
-        bitcoin: {
-          bitcoinPayer: createStubBitcoinPayer(),
-          network: createStubNetwork(),
-          sbtcClient: {} as any,
-          signBitcoinPsbt: () => ({}) as any,
-          broadcast: () => Promise.resolve('test-txid'),
-        },
-      },
-      ...rest,
-    })
-  );
-  return result;
-}
-
-describe('useSwapState', () => {
-  describe('initialization', () => {
-    it('initializes correctly without assets', () => {
-      const result = renderUseSwapState();
-      expect(result.current.state).toEqual({
-        inputCurrencyMode: 'crypto',
-        baseSwapAsset: null,
-        targetSwapAsset: null,
-        selectingAsset: null,
-        assetFlippingAllowed: false,
-        isSendingMax: false,
-        pairReconciliation: { base: 'pending', target: 'pending' },
-        quoteCurrencyPreference: 'USD',
-        quoteStrategy: 'best',
-        baseAmount: '0',
-        secondaryAmount: { status: 'idle', value: null },
-        nonceOverride: undefined,
-        effectiveNonce: 0,
-        slippage: 0.03,
-        customFee: null,
-        feeTier: 'standard',
-      });
-    });
-
-    it('initializes correctly with initial base asset', () => {
-      const result = renderUseSwapState({
-        baseAsset: defaultBtcAsset,
-      });
-
-      expect(result.current.state).toMatchObject({
-        baseSwapAsset: { asset: defaultBtcAsset, providerAssets: [] },
-      });
-    });
-
-    it('initializes correctly with both initial assets', () => {
-      const result = renderUseSwapState({
-        baseAsset: defaultStxAsset,
-        targetAsset: defaultBtcAsset,
-      });
-
-      expect(result.current.state).toMatchObject({
-        baseSwapAsset: { asset: defaultStxAsset, providerAssets: [] },
-        targetSwapAsset: { asset: defaultBtcAsset, providerAssets: [] },
-      });
-    });
-  });
-
-  describe('asset listing', () => {
-    it('sorts base assets with BTC, STX, sBTC priority, then by balance descending', async () => {
-      const baseSwapAssets = [
-        createAccountSwapAsset({
-          asset: { protocol: 'sip10', symbol: 'TOKEN1' },
-          balance: { quote: 500, crypto: 10 },
-        }),
-        createAccountSwapAsset({
-          asset: defaultStxAsset,
-          balance: { quote: 0, crypto: 0 },
-        }),
-        createAccountSwapAsset({
-          asset: { protocol: 'sip10', symbol: 'TOKEN2' },
-          balance: { quote: 1000, crypto: 20 },
-        }),
-        createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { quote: 100, crypto: 0.001 },
-        }),
-        createAccountSwapAsset({
-          asset: defaultSbtcAsset,
-          balance: { quote: 50, crypto: 0.0005 },
-        }),
-      ];
-
-      const result = renderUseSwapState({ baseSwapAssets });
-
-      await waitFor(() => {
-        expect(result.current.baseAssetsQuery.status).toBe('success');
-      });
-
-      const sortedAssets = result.current.baseAssetsQuery.data;
-      assert(sortedAssets);
-      expect(sortedAssets.map(a => a.asset.symbol)).toEqual([
-        'BTC',
-        'STX',
-        'sBTC',
-        'TOKEN2',
-        'TOKEN1',
-      ]);
-    });
-
-    it('filters out assets with zero crypto balance except BTC and STX', async () => {
-      const baseSwapAssets = [
-        createAccountSwapAsset({
-          asset: { protocol: 'sip10', symbol: 'ZERO_TOKEN' },
-          balance: { quote: 0, crypto: 0 },
-        }),
-        createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { quote: 0, crypto: 0 },
-        }),
-        createAccountSwapAsset({
-          asset: defaultStxAsset,
-          balance: { quote: 0, crypto: 0 },
-        }),
-        createAccountSwapAsset({
-          asset: { protocol: 'sip10', symbol: 'ACTIVE_TOKEN' },
-          balance: { quote: 100, crypto: 50 },
-        }),
-      ];
-
-      const result = renderUseSwapState({ baseSwapAssets });
-
-      await waitFor(() => {
-        expect(result.current.baseAssetsQuery.data).toBeDefined();
-      });
-
-      const filteredAssets = result.current.baseAssetsQuery.data;
-      assert(filteredAssets);
-      expect(filteredAssets).toHaveLength(3);
-      expect(filteredAssets.map(a => a.asset.symbol)).toEqual(['BTC', 'STX', 'ACTIVE_TOKEN']);
-      expect(filteredAssets.find(a => a.asset.symbol === 'ZERO_TOKEN')).toBeUndefined();
-    });
-
-    it('sorts target assets with BTC, STX, sBTC priority, then by balance descending', async () => {
-      const targetSwapAssets = [
-        createAccountSwapAsset({
-          asset: { protocol: 'sip10', symbol: 'TOKEN1' },
-          balance: { quote: 500, crypto: 10 },
-        }),
-        createAccountSwapAsset({
-          asset: defaultStxAsset,
-          balance: { quote: 0, crypto: 0 },
-        }),
-        createAccountSwapAsset({
-          asset: { protocol: 'sip10', symbol: 'TOKEN2' },
-          balance: { quote: 1000, crypto: 20 },
-        }),
-        createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { quote: 100, crypto: 0.001 },
-        }),
-        createAccountSwapAsset({
-          asset: defaultSbtcAsset,
-          balance: { quote: 50, crypto: 0.0005 },
-        }),
-      ];
-
-      const result = renderUseSwapState({
-        targetSwapAssets,
-        baseAsset: defaultBtcAsset,
-      });
-
-      await waitFor(() => {
-        expect(result.current.targetAssetsQuery.status).toBe('success');
-      });
-
-      const sortedAssets = result.current.targetAssetsQuery.data;
-      assert(sortedAssets);
-      expect(sortedAssets.map(a => a.asset.symbol)).toEqual([
-        'BTC',
-        'STX',
-        'sBTC',
-        'TOKEN2',
-        'TOKEN1',
-      ]);
-    });
-
-    it('includes all target assets regardless of balance', async () => {
-      const targetSwapAssets = [
-        createAccountSwapAsset({
-          asset: { protocol: 'sip10', symbol: 'ZERO_TOKEN' },
-          balance: { quote: 0, crypto: 0 },
-        }),
-        createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { quote: 0, crypto: 0 },
-        }),
-        createAccountSwapAsset({
-          asset: defaultStxAsset,
-          balance: { quote: 0, crypto: 0 },
-        }),
-        createAccountSwapAsset({
-          asset: { protocol: 'sip10', symbol: 'ACTIVE_TOKEN' },
-          balance: { quote: 100, crypto: 50 },
-        }),
-      ];
-
-      const result = renderUseSwapState({
-        targetSwapAssets,
-        baseAsset: defaultBtcAsset,
-      });
-
-      await waitFor(() => {
-        expect(result.current.targetAssetsQuery.data).toBeDefined();
-      });
-
-      const allAssets = result.current.targetAssetsQuery.data;
-      assert(allAssets);
-      expect(allAssets).toHaveLength(4);
-      expect(allAssets.map(a => a.asset.symbol)).toEqual([
-        'BTC',
-        'STX',
-        'ACTIVE_TOKEN',
-        'ZERO_TOKEN',
-      ]);
-    });
-  });
-
-  describe('asset selection state', () => {
-    it('sets state to base when opening base selection', () => {
-      const result = renderUseSwapState();
-
-      expect(result.current.state.selectingAsset).toBeNull();
-      act(() => result.current.actions.openAssetSelector('base'));
-      expect(result.current.state.selectingAsset).toBe('base');
-    });
-
-    it('sets state to target when opening target selection', () => {
-      const result = renderUseSwapState();
-
-      expect(result.current.state.selectingAsset).toBeNull();
-      act(() => result.current.actions.openAssetSelector('target'));
-      expect(result.current.state.selectingAsset).toBe('target');
-    });
-
-    it('sets state to null when closing selection', () => {
-      const result = renderUseSwapState();
-
-      act(() => result.current.actions.openAssetSelector('base'));
-      expect(result.current.state.selectingAsset).toBe('base');
-
-      act(() => result.current.actions.closeAssetSelector());
-      expect(result.current.state.selectingAsset).toBeNull();
-    });
-  });
-
-  describe('base asset selection', () => {
-    it('updates state immediately when setting base asset', () => {
-      const result = renderUseSwapState();
-
-      const newAsset = createAccountSwapAsset({
-        asset: { protocol: 'sip10', symbol: 'TEST_TOKEN' },
-        balance: { crypto: 50, quote: 100 },
-      });
-
-      act(() => result.current.actions.setBaseSwapAsset(newAsset));
-      expect(result.current.state.baseSwapAsset).toEqual(newAsset);
-    });
-
-    it('preserves existing amount value when setting base asset', () => {
-      const result = renderUseSwapState({
-        baseAsset: defaultBtcAsset,
-      });
-
-      act(() => result.current.actions.setBaseAmount('123.45'));
-
-      const newAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 50, quote: 100 },
-      });
-
-      act(() => result.current.actions.setBaseSwapAsset(newAsset));
-      expect(result.current.state.baseAmount).toBe('123.45');
-    });
-
-    it('adjusts decimal places to match new asset decimals in crypto input mode', () => {
-      const result = renderUseSwapState({
-        baseAsset: defaultBtcAsset,
-      });
-
-      expect(result.current.state.inputCurrencyMode).toBe('crypto');
-
-      act(() => result.current.actions.setBaseAmount('123.45678912'));
-
-      const twoDecimalAsset = createAccountSwapAsset({
-        asset: { protocol: 'sip10', symbol: 'TWO_DECIMAL', decimals: 2 },
-        balance: { crypto: 50, quote: 100 },
-      });
-
-      act(() => result.current.actions.setBaseSwapAsset(twoDecimalAsset));
-      expect(result.current.state.baseAmount).toBe('123.45');
-    });
-
-    it('resets asset selection state after base asset is set', () => {
-      const result = renderUseSwapState();
-
-      act(() => result.current.actions.openAssetSelector('base'));
-      expect(result.current.state.selectingAsset).toBe('base');
-
-      const newAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 50, quote: 100 },
-      });
-
-      act(() => result.current.actions.setBaseSwapAsset(newAsset));
-      expect(result.current.state.selectingAsset).toBeNull();
-    });
-  });
-
-  describe('target asset selection', () => {
-    it('updates state immediately when setting target asset', () => {
-      const result = renderUseSwapState({
-        baseAsset: defaultBtcAsset,
-      });
-
-      const newTargetAsset = createAccountSwapAsset({
-        asset: { protocol: 'sip10', symbol: 'TARGET_TOKEN' },
-        balance: { crypto: 25, quote: 200 },
-      });
-
-      act(() => result.current.actions.setTargetSwapAsset(newTargetAsset));
-      expect(result.current.state.targetSwapAsset).toEqual(newTargetAsset);
-    });
-
-    it('resets asset selection state after target asset is set', () => {
-      const result = renderUseSwapState({
-        baseAsset: defaultBtcAsset,
-      });
-
-      act(() => result.current.actions.openAssetSelector('target'));
-      expect(result.current.state.selectingAsset).toBe('target');
-
-      const newTargetAsset = createAccountSwapAsset({
-        asset: defaultSbtcAsset,
-        balance: { crypto: 10, quote: 500 },
-      });
-
-      act(() => result.current.actions.setTargetSwapAsset(newTargetAsset));
-      expect(result.current.state.selectingAsset).toBeNull();
-    });
-  });
-
-  describe('initial placholder base asset reconciliation', () => {
-    it('replaces the initial asset with a real asset when data arrives', async () => {
-      const realSwapAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 1000, crypto: 0.5 },
-      });
-
-      const result = renderUseSwapState({
-        baseAsset: defaultBtcAsset,
-        baseSwapAssets: [realSwapAsset],
-      });
-
-      expect(result.current.state.baseSwapAsset).toEqual({
-        asset: defaultBtcAsset,
-        providerAssets: [],
-      });
-      expect(result.current.state.pairReconciliation.base).toBe('pending');
-
-      await waitFor(() => {
-        expect(result.current.state.pairReconciliation.base).toBe('complete');
-      });
-
-      expect(result.current.state.baseSwapAsset).toEqual(realSwapAsset);
-    });
-
-    it('clears both assets and resets the amount when initial base asset not found in data', async () => {
-      const differentAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { quote: 500, crypto: 100 },
-      });
-
-      const result = renderUseSwapState({
-        baseAsset: defaultBtcAsset,
-        targetAsset: defaultStxAsset,
-        baseSwapAssets: [differentAsset],
-      });
-
-      act(() => result.current.actions.setBaseAmount('123.45'));
-
-      expect(result.current.state.baseSwapAsset).toBeTruthy();
-      expect(result.current.state.targetSwapAsset).toBeTruthy();
-      expect(result.current.state.baseAmount).toBe('123.45');
-
-      await waitFor(() => {
-        expect(result.current.baseAssetsQuery.status).toBe('success');
-      });
-
-      expect(result.current.state.baseSwapAsset).toBeNull();
-      expect(result.current.state.targetSwapAsset).toBeNull();
-      expect(result.current.state.baseAmount).toBe('0');
-      expect(result.current.state.pairReconciliation).toMatchObject({
-        base: 'pending',
-      });
-    });
-  });
-
-  describe('initial placholder target asset reconciliation', () => {
-    it('replaces the initial asset with a real asset when data arrives', async () => {
-      const realBaseAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 1000, crypto: 0.5 },
-      });
-
-      const realTargetAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { quote: 500, crypto: 100 },
-      });
-
-      const result = renderUseSwapState({
-        baseAsset: defaultBtcAsset,
-        targetAsset: defaultStxAsset,
-        baseSwapAssets: [realBaseAsset],
-        targetSwapAssets: [realTargetAsset],
-      });
-
-      expect(result.current.state.targetSwapAsset).toEqual({
-        asset: defaultStxAsset,
-        providerAssets: [],
-      });
-      expect(result.current.state.pairReconciliation.target).toBe('pending');
-
-      await waitFor(() => {
-        expect(result.current.state.pairReconciliation.target).toBe('complete');
-      });
-
-      expect(result.current.state.targetSwapAsset).toEqual(realTargetAsset);
-    });
-
-    it('clears target asset when initial target asset not found in data', async () => {
-      const realBaseAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 1000, crypto: 0.5 },
-      });
-
-      const differentTargetAsset = createAccountSwapAsset({
-        asset: defaultSbtcAsset,
-        balance: { quote: 200, crypto: 0.002 },
-      });
-
-      const result = renderUseSwapState({
-        baseAsset: defaultBtcAsset,
-        targetAsset: defaultStxAsset,
-        baseSwapAssets: [realBaseAsset],
-        targetSwapAssets: [differentTargetAsset],
-      });
-
-      expect(result.current.state.targetSwapAsset).toBeTruthy();
-      expect(result.current.state.baseSwapAsset).toBeTruthy();
-
-      await waitFor(() => {
-        expect(result.current.targetAssetsQuery.status).toBe('success');
-      });
-
-      expect(result.current.state.targetSwapAsset).toBeNull();
-      expect(result.current.state.baseSwapAsset).toEqual(realBaseAsset);
-      expect(result.current.state.pairReconciliation.target).toBe('pending');
-    });
-
-    it('reconciles target asset when base asset changes', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 1000, crypto: 0.5 },
-      });
-
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { quote: 500, crypto: 100 },
-      });
-
-      const sbtcAsset = createAccountSwapAsset({
-        asset: defaultSbtcAsset,
-        balance: { quote: 200, crypto: 0.002 },
-      });
-
-      const result = renderUseSwapState({
-        baseAsset: defaultStxAsset,
-        targetAsset: defaultSbtcAsset,
-        baseSwapAssets: [btcAsset, stxAsset],
-        targetSwapAssets: [sbtcAsset],
-      });
-
-      await waitFor(() => {
-        expect(result.current.state.pairReconciliation.base).toBe('complete');
-        expect(result.current.state.pairReconciliation.target).toBe('complete');
-      });
-
-      expect(result.current.state.targetSwapAsset).toEqual(sbtcAsset);
-
-      act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-
-      expect(result.current.state.pairReconciliation.target).toBe('pending');
-
-      await waitFor(() => {
-        expect(result.current.state.pairReconciliation.target).toBe('complete');
-      });
-
-      expect(result.current.state.targetSwapAsset).toEqual(sbtcAsset);
-    });
-  });
-
-  describe('flipping base and target assets', () => {
-    it('flips base and target positions', () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 1000, crypto: 0.5 },
-      });
-
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { quote: 500, crypto: 100 },
-      });
-
-      const result = renderUseSwapState();
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-      });
-
-      expect(result.current.state.baseSwapAsset).toEqual(btcAsset);
-      expect(result.current.state.targetSwapAsset).toEqual(stxAsset);
-
-      act(() => result.current.actions.flipAssets());
-
-      expect(result.current.state.baseSwapAsset).toEqual(stxAsset);
-      expect(result.current.state.targetSwapAsset).toEqual(btcAsset);
-    });
-
-    it('only flips if both assets are set', () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 1000, crypto: 0.5 },
-      });
-
-      const result = renderUseSwapState();
-
-      act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-      expect(result.current.state.baseSwapAsset).toEqual(btcAsset);
-      expect(result.current.state.targetSwapAsset).toBeNull();
-
-      act(() => result.current.actions.flipAssets());
-      expect(result.current.state.baseSwapAsset).toEqual(btcAsset);
-      expect(result.current.state.targetSwapAsset).toBeNull();
-    });
-
-    it('resets base amount to 0 when flipping', () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 1000, crypto: 0.5 },
-      });
-
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { quote: 500, crypto: 100 },
-      });
-
-      const result = renderUseSwapState();
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-        result.current.actions.setBaseAmount('123.45');
-      });
-      expect(result.current.state.baseAmount).toBe('123.45');
-      expect(result.current.state.isSendingMax).toBe(false);
-
-      act(() => result.current.actions.flipAssets());
-      expect(result.current.state.baseAmount).toBe('0');
-    });
-
-    it('indicates asset flipping is allowed', () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 1000, crypto: 0.5 },
-      });
-
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { quote: 500, crypto: 100 },
-      });
-
-      const result = renderUseSwapState();
-
-      expect(result.current.state.assetFlippingAllowed).toBe(false);
-
-      act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-      expect(result.current.state.assetFlippingAllowed).toBe(false);
-
-      act(() => result.current.actions.setTargetSwapAsset(stxAsset));
-      expect(result.current.state.assetFlippingAllowed).toBe(true);
-
-      act(() => result.current.actions.clearAssetSelection());
-      expect(result.current.state.assetFlippingAllowed).toBe(false);
-    });
-
-    it('marks both assets for re-reconciliation', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 1000, crypto: 0.5 },
-      });
-
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { quote: 500, crypto: 100 },
-      });
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset, stxAsset],
-        targetSwapAssets: [btcAsset, stxAsset],
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-      });
-      await waitFor(() => {
-        expect(result.current.state.pairReconciliation.base).toBe('complete');
-        expect(result.current.state.pairReconciliation.target).toBe('complete');
-      });
-
-      act(() => result.current.actions.flipAssets());
-      expect(result.current.state.pairReconciliation.base).toBe('pending');
-      expect(result.current.state.pairReconciliation.target).toBe('pending');
-    });
-
-    it('flips automatically when selecting current target as new base', () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 1000, crypto: 0.5 },
-      });
-
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { quote: 500, crypto: 100 },
-      });
-
-      const result = renderUseSwapState({
-        baseAsset: defaultBtcAsset,
-        targetAsset: defaultStxAsset,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-        result.current.actions.setBaseAmount('123.45');
-      });
-
-      expect(result.current.state.baseSwapAsset).toEqual(btcAsset);
-      expect(result.current.state.targetSwapAsset).toEqual(stxAsset);
-      expect(result.current.state.baseAmount).toBe('123.45');
-
-      act(() => result.current.actions.setBaseSwapAsset(stxAsset));
-
-      expect(result.current.state.baseSwapAsset).toEqual(stxAsset);
-      expect(result.current.state.targetSwapAsset).toEqual(btcAsset);
-      expect(result.current.state.baseAmount).toBe('0');
-      expect(result.current.state.pairReconciliation.base).toBe('pending');
-      expect(result.current.state.pairReconciliation.target).toBe('pending');
-      expect(result.current.state.selectingAsset).toBeNull();
-    });
-  });
-  describe('amount presets', () => {
-    describe('setting preset percentages', () => {
-      it('sets base amount to 25%, 50%, 75%, and 100% of available balance in crypto mode', () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-        });
-
-        act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-        expect(result.current.state.inputCurrencyMode).toBe('crypto');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.25));
-        expect(result.current.state.baseAmount).toBe('0.25');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.5));
-        expect(result.current.state.baseAmount).toBe('0.5');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.75));
-        expect(result.current.state.baseAmount).toBe('0.75');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.baseAmount).toBe('1');
-      });
-
-      it('sets base amount to 25%, 50%, 75%, and 100% of available balance in quote mode', () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-        });
-
-        act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-        act(() => result.current.actions.toggleInputCurrencyMode());
-        expect(result.current.state.inputCurrencyMode).toBe('quote');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.25));
-        expect(result.current.state.baseAmount).toBe('12500');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.5));
-        expect(result.current.state.baseAmount).toBe('25000');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.75));
-        expect(result.current.state.baseAmount).toBe('37500');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.baseAmount).toBe('50000');
-      });
-    });
-
-    describe('preset edge cases', () => {
-      it('returns early when no base asset is selected', () => {
-        const result = renderUseSwapState();
-        expect(result.current.state.baseSwapAsset).toBeNull();
-        act(() => result.current.actions.setBaseAmountByPercentage(0.5));
-        expect(result.current.state.baseAmount).toBe('0');
-      });
-
-      it('returns early when base asset has no balance', () => {
-        const assetWithoutBalance = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-        });
-        const result = renderUseSwapState();
-        act(() => result.current.actions.setBaseSwapAsset(assetWithoutBalance));
-        expect(result.current.state.baseSwapAsset?.balance).toBeUndefined();
-        act(() => result.current.actions.setBaseAmountByPercentage(0.5));
-        expect(result.current.state.baseAmount).toBe('0');
-      });
-
-      it('handles zero available balance correctly', () => {
-        const zeroBalanceAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 0, quote: 0 },
-        });
-        const result = renderUseSwapState();
-        act(() => result.current.actions.setBaseSwapAsset(zeroBalanceAsset));
-        act(() => result.current.actions.setBaseAmountByPercentage(0.5));
-        expect(result.current.state.baseAmount).toBe('0');
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.baseAmount).toBe('0');
-      });
-
-      it('preserves precision with very small balances', () => {
-        const smallBalanceAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100, quote: 5 },
-        });
-        const result = renderUseSwapState();
-        act(() => result.current.actions.setBaseSwapAsset(smallBalanceAsset));
-        act(() => result.current.actions.setBaseAmountByPercentage(0.5));
-        expect(result.current.state.baseAmount).toBe('0.0000005');
-        act(() => result.current.actions.setBaseAmountByPercentage(0.25));
-        expect(result.current.state.baseAmount).toBe('0.00000025');
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.baseAmount).toBe('0.000001');
-      });
-
-      it('handles very large balances without overflow', () => {
-        const largeBalanceAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 2100000000000000, quote: 105000000000000 },
-        });
-        const result = renderUseSwapState();
-        act(() => result.current.actions.setBaseSwapAsset(largeBalanceAsset));
-        act(() => result.current.actions.setBaseAmountByPercentage(0.25));
-        expect(result.current.state.baseAmount).toBe('5250000');
-        act(() => result.current.actions.setBaseAmountByPercentage(0.5));
-        expect(result.current.state.baseAmount).toBe('10500000');
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.baseAmount).toBe('21000000');
-      });
-
-      it('maintains precision with very small percentage calculations', () => {
-        const precisionAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 546, quote: 6 },
-        });
-        const result = renderUseSwapState({
-          baseSwapAssets: [precisionAsset],
-        });
-        act(() => result.current.actions.setBaseSwapAsset(precisionAsset));
-        act(() => result.current.actions.setBaseAmountByPercentage(0.25));
-
-        const expectedValue = 0.00000546 * 0.25;
-        expect(parseFloat(result.current.state.baseAmount)).toBeCloseTo(expectedValue, 7);
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.baseAmount).toBe('0.00000546');
-        expect(result.current.state.isSendingMax).toBe(true);
-      });
-    });
-
-    describe('presets with currency mode switching', () => {
-      it('calculates preset correctly after toggling from crypto to quote mode', async () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-        const marketData = {
-          pair: { base: 'BTC', quote: 'USD' },
-          price: createMoney(50_000_00, 'USD', 2),
-        };
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-          marketData,
-        });
-
-        act(() => {
-          result.current.actions.setBaseSwapAsset(btcAsset);
-          result.current.actions.setBaseAmount('0.1');
-        });
-        await waitFor(() => expect(result.current.state.secondaryAmount.status).toBe('success'));
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.5));
-        expect(result.current.state.baseAmount).toBe('0.5');
-
-        act(() => result.current.actions.toggleInputCurrencyMode());
-        expect(result.current.state.inputCurrencyMode).toBe('quote');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.25));
-        expect(result.current.state.baseAmount).toBe('12500');
-      });
-
-      it('calculates preset correctly after toggling from quote to crypto mode', async () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-        const marketData = {
-          pair: { base: 'BTC', quote: 'USD' },
-          price: createMoney(50_000_00, 'USD', 2),
-        };
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-          marketData,
-        });
-        act(() => {
-          result.current.actions.setBaseSwapAsset(btcAsset);
-          result.current.actions.setBaseAmount('0.1');
-        });
-        await waitFor(() => expect(result.current.state.secondaryAmount.status).toBe('success'));
-
-        act(() => result.current.actions.toggleInputCurrencyMode());
-        expect(result.current.state.inputCurrencyMode).toBe('quote');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.5));
-        expect(result.current.state.baseAmount).toBe('25000');
-
-        act(() => result.current.actions.toggleInputCurrencyMode());
-        expect(result.current.state.inputCurrencyMode).toBe('crypto');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.75));
-        expect(result.current.state.baseAmount).toBe('0.75');
-      });
-
-      it('respects the current input currency mode when calculating preset', async () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-        const marketData = {
-          pair: { base: 'BTC', quote: 'USD' },
-          price: createMoney(50_000_00, 'USD', 2),
-        };
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-          marketData,
-        });
-
-        act(() => {
-          result.current.actions.setBaseSwapAsset(btcAsset);
-          result.current.actions.setBaseAmount('0.1');
-        });
-        await waitFor(() => expect(result.current.state.secondaryAmount.status).toBe('success'));
-        expect(result.current.state.inputCurrencyMode).toBe('crypto');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.baseAmount).toBe('1');
-
-        act(() => result.current.actions.toggleInputCurrencyMode());
-        expect(result.current.state.inputCurrencyMode).toBe('quote');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.baseAmount).toBe('50000');
-      });
-    });
-  });
-
-  describe('isSendingMax flag', () => {
-    describe('basic isSendingMax detection', () => {
-      it('returns true when amount equals 100% in crypto mode', () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-        });
-
-        act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-        expect(result.current.state.inputCurrencyMode).toBe('crypto');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.isSendingMax).toBe(true);
-      });
-
-      it('returns true when amount equals 100% in quote mode', async () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-        const marketData = {
-          pair: { base: 'BTC', quote: 'USD' },
-          price: createMoney(50_000_00, 'USD', 2),
-        };
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-          marketData,
-        });
-        act(() => {
-          result.current.actions.setBaseSwapAsset(btcAsset);
-          result.current.actions.setBaseAmount('0.1');
-        });
-        await waitFor(() => expect(result.current.state.secondaryAmount.status).toBe('success'));
-
-        act(() => result.current.actions.toggleInputCurrencyMode());
-        expect(result.current.state.inputCurrencyMode).toBe('quote');
-
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.isSendingMax).toBe(true);
-      });
-
-      it('returns false when amount is less than 100%', () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-        });
-
-        act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-        act(() => result.current.actions.setBaseAmountByPercentage(0.25));
-        expect(result.current.state.isSendingMax).toBe(false);
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.5));
-        expect(result.current.state.isSendingMax).toBe(false);
-
-        act(() => result.current.actions.setBaseAmountByPercentage(0.75));
-        expect(result.current.state.isSendingMax).toBe(false);
-      });
-    });
-
-    describe('isSendingMax with manual input', () => {
-      it('becomes true when manually entering exact available balance in crypto mode', () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 12345678, quote: 617284 },
-        });
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-        });
-
-        act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-        expect(result.current.state.inputCurrencyMode).toBe('crypto');
-
-        act(() => result.current.actions.setBaseAmount('0.12345678'));
-        expect(result.current.state.isSendingMax).toBe(true);
-      });
-
-      it('becomes true when manually entering exact available balance in quote mode', async () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-        const marketData = {
-          pair: { base: 'BTC', quote: 'USD' },
-          price: createMoney(50_000_00, 'USD', 2),
-        };
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-          marketData,
-        });
-        act(() => {
-          result.current.actions.setBaseSwapAsset(btcAsset);
-          result.current.actions.setBaseAmount('0.1');
-        });
-        await waitFor(() => expect(result.current.state.secondaryAmount.status).toBe('success'));
-        act(() => result.current.actions.toggleInputCurrencyMode());
-        expect(result.current.state.inputCurrencyMode).toBe('quote');
-
-        act(() => result.current.actions.setBaseAmount('50000'));
-        expect(result.current.state.isSendingMax).toBe(true);
-      });
-
-      it('becomes false when editing amount down from max', () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-        });
-
-        act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.isSendingMax).toBe(true);
-
-        act(() => result.current.actions.setBaseAmount('0.99999'));
-        expect(result.current.state.isSendingMax).toBe(false);
-      });
-
-      it('remains accurate after toggling currency modes', async () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-        const marketData = {
-          pair: { base: 'BTC', quote: 'USD' },
-          price: createMoney(50_000_00, 'USD', 2),
-        };
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset],
-          marketData,
-        });
-        act(() => {
-          result.current.actions.setBaseSwapAsset(btcAsset);
-          result.current.actions.setBaseAmount('0.1');
-        });
-        await waitFor(() => expect(result.current.state.secondaryAmount.status).toBe('success'));
-
-        act(() => result.current.actions.setBaseAmountByPercentage(1));
-        expect(result.current.state.isSendingMax).toBe(true);
-
-        act(() => result.current.actions.toggleInputCurrencyMode());
-        expect(result.current.state.isSendingMax).toBe(true);
-      });
-    });
-  });
-
-  describe('preset and isSendingMax integration', () => {
-    it('sets isSendingMax to true after selecting MAX preset', () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-      });
-
-      act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-      act(() => result.current.actions.setBaseAmountByPercentage(1));
-      expect(result.current.state.baseAmount).toBe('1');
-      expect(result.current.state.isSendingMax).toBe(true);
-    });
-
-    it('sets isSendingMax to false after selecting a preset', () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-      });
-      act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-      act(() => result.current.actions.setBaseAmountByPercentage(0.25));
-      expect(result.current.state.isSendingMax).toBe(false);
-    });
-
-    it('updates isSendingMax when switching between assets', () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 100000000, quote: 15000 },
-      });
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset, stxAsset],
-      });
-      act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-      act(() => result.current.actions.setBaseAmountByPercentage(1));
-      expect(result.current.state.isSendingMax).toBe(true);
-
-      //act(() => result.current.actions.setBaseSwapAsset(stxAsset));
-      //expect(result.current.state.isSendingMax).toBe(false);
-    });
-  });
-
-  describe('quote currency calculation', () => {
-    it('calculates entered amount value in selected quote currency', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 50000, crypto: 1 },
-      });
-
-      const marketData = {
-        pair: { base: 'BTC', quote: 'USD' },
-        price: createMoney(100_000_00, 'USD', 2),
-      };
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        marketData,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-      expect(result.current.state.inputCurrencyMode).toBe('crypto');
-      expect(result.current.state.baseAmount).toBe('0.1');
-
-      await waitFor(() => expect(result.current.state.secondaryAmount.status).toBe('success'));
-      expect(result.current.state.secondaryAmount.value?.amount.toNumber()).toBe(10_000_00);
-    });
-
-    it('switches to quote mode when base amount is set', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 50000, crypto: 1 },
-      });
-
-      const marketData = {
-        pair: { base: 'BTC', quote: 'USD' },
-        price: createMoney(100_000_00, 'USD', 2),
-      };
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        marketData,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-      await waitFor(() => expect(result.current.state.secondaryAmount.status).toBe('success'));
-      expect(result.current.state.inputCurrencyMode).toBe('crypto');
-      expect(result.current.state.baseAmount).toBe('0.1');
-      expect(result.current.state.secondaryAmount.value?.amount.toNumber()).toBe(10_000_00);
-
-      act(() => result.current.actions.toggleInputCurrencyMode());
-      expect(result.current.state.inputCurrencyMode).toBe('quote');
-      expect(result.current.state.baseAmount).toBe('10000');
-      expect(result.current.state.secondaryAmount.value?.amount.toNumber()).toBe(10_000_000);
-    });
-
-    it('preserves exact values when toggled twice', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 50000, crypto: 1 },
-      });
-
-      const marketData = {
-        pair: { base: 'BTC', quote: 'USD' },
-        price: createMoney(100_000_00, 'USD', 2),
-      };
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        marketData,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-
-      await waitFor(() => expect(result.current.state.secondaryAmount.status).toBe('success'));
-
-      act(() => result.current.actions.toggleInputCurrencyMode());
-      expect(result.current.state.inputCurrencyMode).toBe('quote');
-      expect(result.current.state.baseAmount).toBe('10000');
-
-      act(() => result.current.actions.toggleInputCurrencyMode());
-      expect(result.current.state.inputCurrencyMode).toBe('crypto');
-      expect(result.current.state.baseAmount).toBe('0.1');
-      expect(result.current.state.secondaryAmount.value?.amount.toNumber()).toBe(10_000_00);
-    });
-    it('returns idle state when no base asset is selected', () => {
-      const result = renderUseSwapState({});
-
-      expect(result.current.state.secondaryAmount.status).toBe('idle');
-      expect(result.current.state.secondaryAmount.value).toBeNull();
-    });
-
-    it('returns pending state while market data is loading', () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 50000, crypto: 1 },
-      });
-
-      const marketData = {
-        pair: { base: 'BTC', quote: 'USD' },
-        price: createMoney(100_000_00, 'USD', 2),
-      };
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        marketData,
-      });
-
-      act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-      expect(result.current.state.secondaryAmount.status).toBe('pending');
-      expect(result.current.state.secondaryAmount.value).toBeNull();
-    });
-
-    it('returns error state when market data fails', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { quote: 50000, crypto: 1 },
-      });
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-      });
-
-      act(() => result.current.actions.setBaseSwapAsset(btcAsset));
-      await waitFor(() => expect(result.current.state.secondaryAmount.status).toBe('error'));
-      expect(result.current.state.secondaryAmount.value).toBeNull();
-    });
-  });
-
-  describe('quote enrichment and selection', () => {
-    it('initializes with quoteStrategy set to best', () => {
-      const result = renderUseSwapState();
-      expect(result.current.state.quoteStrategy).toBe('best');
-    });
-
-    it('returns quotes and selected structure from quoteQuery', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 500_000_000, quote: 5_000_00 },
-      });
-
-      const quotes = [
-        createSwapQuote({ targetAmount: 600_000_000, providerId: 'alex-sdk' }),
-        createSwapQuote({ targetAmount: 550_000_000, providerId: 'velar-sdk' }),
-        createSwapQuote({ targetAmount: 500_000_000, providerId: 'bitflow-sdk' }),
-      ];
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        targetSwapAssets: [stxAsset],
-        swapQuotes: quotes,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-
-      await waitFor(() => {
-        expect(result.current.quoteQuery.data).toBeDefined();
-      });
-
-      const quoteData = result.current.quoteQuery.data;
-      assert(quoteData);
-      expect(quoteData.quotes).toHaveLength(3);
-      expect(quoteData.selected).toBeDefined();
-      expect(quoteData.selected?.rawSwapQuote.targetAmount).toBe(600_000_000);
-    });
-
-    it('selects the quote with highest target amount as best', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 500_000_000, quote: 5_000_00 },
-      });
-
-      const quotes = [
-        createSwapQuote({ targetAmount: 400_000_000, providerId: 'bitflow-sdk' }),
-        createSwapQuote({ targetAmount: 700_000_000, providerId: 'alex-sdk' }),
-        createSwapQuote({ targetAmount: 550_000_000, providerId: 'velar-sdk' }),
-      ];
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        targetSwapAssets: [stxAsset],
-        swapQuotes: quotes,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-
-      await waitFor(() => {
-        expect(result.current.quoteQuery.data).toBeDefined();
-      });
-
-      const selected = result.current.quoteQuery.data?.selected;
-      expect(selected?.rawSwapQuote.targetAmount).toBe(700_000_000);
-      expect(selected?.provider).toBe('alex-sdk');
-    });
-
-    it('enriches quotes with rate, score, and provider information', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 500_000_000, quote: 5_000_00 },
-      });
-
-      const quotes = [
-        createSwapQuote({
-          baseAmount: 100_000_000,
-          targetAmount: 500_000_000,
-          providerId: 'alex-sdk',
-          dexPath: [
-            {
-              name: 'AlexLab',
-              url: 'https://alexlab.co',
-              logo: 'https://alexlab.co/logo.png',
-              description: 'AlexLab DEX',
-            },
-          ],
-        }),
-      ];
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        targetSwapAssets: [stxAsset],
-        swapQuotes: quotes,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-
-      await waitFor(() => {
-        expect(result.current.quoteQuery.data).toBeDefined();
-      });
-
-      const enrichedQuote = result.current.quoteQuery.data?.quotes[0];
-      assert(enrichedQuote);
-      expect(enrichedQuote.rate).toBe(5);
-      expect(enrichedQuote.score).toBe(500_000_000);
-      expect(enrichedQuote.provider).toBe('alex-sdk');
-      expect(enrichedQuote.quoteAmount).toBeDefined();
-      expect(enrichedQuote.dexPath).toHaveLength(1);
-      expect(enrichedQuote.dexPath[0]).toHaveProperty('name', 'AlexLab');
-      expect(enrichedQuote.assetPath).toHaveLength(2);
-    });
-
-    it('returns empty quotes and undefined selected when no quotes available', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 500_000_000, quote: 5_000_00 },
-      });
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        targetSwapAssets: [stxAsset],
-        swapQuotes: [],
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-
-      await waitFor(() => {
-        expect(result.current.quoteQuery.data).toBeDefined();
-      });
-
-      expect(result.current.quoteQuery.data?.quotes).toEqual([]);
-      expect(result.current.quoteQuery.data?.selected).toBeUndefined();
-    });
-
-    it('returns single quote as selected when only one quote available', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 500_000_000, quote: 5_000_00 },
-      });
-
-      const singleQuote = createSwapQuote({
-        targetAmount: 450_000_000,
-        providerId: 'velar-sdk',
-      });
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        targetSwapAssets: [stxAsset],
-        swapQuotes: [singleQuote],
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-
-      await waitFor(() => {
-        expect(result.current.quoteQuery.data).toBeDefined();
-      });
-
-      expect(result.current.quoteQuery.data?.quotes).toHaveLength(1);
-      expect(result.current.quoteQuery.data?.selected?.rawSwapQuote).toEqual(singleQuote);
-    });
-
-    it('calculates provider fee based on dex path length for stacks contract calls', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 500_000_000, quote: 5_000_00 },
-      });
-
-      const quotes = [
-        createSwapQuote({
-          executionType: 'stacks-contract-call',
-          targetAmount: 500_000_000,
-          providerId: 'alex-sdk',
-          dexPath: [
-            {
-              name: 'AlexLab',
-              url: 'https://alexlab.co',
-              logo: 'https://alexlab.co/logo.png',
-              description: 'AlexLab DEX',
-            },
-          ],
-        }),
-        createSwapQuote({
-          executionType: 'stacks-contract-call',
-          targetAmount: 450_000_000,
-          providerId: 'velar-sdk',
-          dexPath: [
-            {
-              name: 'Velar',
-              url: 'https://velar.co',
-              logo: 'https://velar.co/logo.png',
-              description: 'Velar DEX',
-            },
-            {
-              name: 'AlexLab',
-              url: 'https://alexlab.co',
-              logo: 'https://alexlab.co/logo.png',
-              description: 'AlexLab DEX',
-            },
-          ],
-        }),
-        createSwapQuote({
-          executionType: 'stacks-contract-call',
-          targetAmount: 400_000_000,
-          providerId: 'bitflow-sdk',
-          dexPath: [
-            {
-              name: 'Bitflow',
-              url: 'https://bitflow.finance',
-              logo: 'https://bitflow.finance/logo.png',
-              description: 'Bitflow DEX',
-            },
-            {
-              name: 'Velar',
-              url: 'https://velar.co',
-              logo: 'https://velar.co/logo.png',
-              description: 'Velar DEX',
-            },
-            {
-              name: 'AlexLab',
-              url: 'https://alexlab.co',
-              logo: 'https://alexlab.co/logo.png',
-              description: 'AlexLab DEX',
-            },
-          ],
-        }),
-      ];
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        targetSwapAssets: [stxAsset],
-        swapQuotes: quotes,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-
-      await waitFor(() => {
-        expect(result.current.quoteQuery.data).toBeDefined();
-      });
-
-      const enrichedQuotes = result.current.quoteQuery.data?.quotes;
-      assert(enrichedQuotes);
-      expect(enrichedQuotes).toHaveLength(3);
-
-      const singleDexQuote = enrichedQuotes.find(q => q.dexPath.length === 1);
-      const doubleDexQuote = enrichedQuotes.find(q => q.dexPath.length === 2);
-      const tripleDexQuote = enrichedQuotes.find(q => q.dexPath.length === 3);
-
-      expect(singleDexQuote?.providerFee).toBeCloseTo(0.003, 5);
-      expect(doubleDexQuote?.providerFee).toBeCloseTo(0.006, 5);
-      expect(tripleDexQuote?.providerFee).toBeCloseTo(0.009, 5);
-    });
-
-    it('does not add provider fee for sbtc bridge transfers', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const sbtcAsset = createAccountSwapAsset({
-        asset: defaultSbtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-
-      const quotes = [
-        createSwapQuote({
-          executionType: 'sbtc-bridge-transfer',
-          targetAmount: 500_000_000,
-          providerId: 'sbtc-bridge',
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultSbtcAsset,
-        }),
-      ];
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        targetSwapAssets: [sbtcAsset],
-        swapQuotes: quotes,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(sbtcAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-
-      await waitFor(() => {
-        expect(result.current.quoteQuery.data).toBeDefined();
-      });
-
-      const bridgeQuote = result.current.quoteQuery.data?.quotes[0];
-      assert(bridgeQuote);
-      expect(bridgeQuote.providerFee).toBeUndefined();
-      expect(bridgeQuote.rawSwapQuote.executionType).toBe('sbtc-bridge-transfer');
-    });
-
-    it('sorts quotes by score in descending order', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 500_000_000, quote: 5_000_00 },
-      });
-
-      const quotes = [
-        createSwapQuote({ targetAmount: 300_000_000, providerId: 'bitflow-sdk' }),
-        createSwapQuote({ targetAmount: 700_000_000, providerId: 'alex-sdk' }),
-        createSwapQuote({ targetAmount: 500_000_000, providerId: 'velar-sdk' }),
-        createSwapQuote({ targetAmount: 900_000_000, providerId: 'alex-sdk' }),
-        createSwapQuote({ targetAmount: 100_000_000, providerId: 'bitflow-sdk' }),
-      ];
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        targetSwapAssets: [stxAsset],
-        swapQuotes: quotes,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-
-      await waitFor(() => {
-        expect(result.current.quoteQuery.data).toBeDefined();
-      });
-
-      const sortedQuotes = result.current.quoteQuery.data?.quotes;
-      assert(sortedQuotes);
-      expect(sortedQuotes).toHaveLength(5);
-
-      for (let i = 1; i < sortedQuotes.length; i++) {
-        expect(sortedQuotes[i - 1]?.score).toBeGreaterThanOrEqual(sortedQuotes[i]?.score ?? 0);
-      }
-
-      expect(sortedQuotes[0]?.score).toBe(900_000_000);
-      expect(sortedQuotes[4]?.score).toBe(100_000_000);
-      expect(result.current.quoteQuery.data?.selected?.score).toBe(900_000_000);
-    });
-
-    it('sets score equal to targetAmount for all quotes', async () => {
-      const btcAsset = createAccountSwapAsset({
-        asset: defaultBtcAsset,
-        balance: { crypto: 100_000_000, quote: 50_000_00 },
-      });
-      const stxAsset = createAccountSwapAsset({
-        asset: defaultStxAsset,
-        balance: { crypto: 500_000_000, quote: 5_000_00 },
-      });
-
-      const quotes = [
-        createSwapQuote({
-          executionType: 'stacks-contract-call',
-          targetAmount: 456_789_000,
-          providerId: 'alex-sdk',
-        }),
-        createSwapQuote({
-          executionType: 'sbtc-bridge-transfer',
-          targetAmount: 123_456_000,
-          providerId: 'sbtc-bridge',
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultSbtcAsset,
-        }),
-        createSwapQuote({
-          executionType: 'stacks-contract-call',
-          targetAmount: 987_654_321,
-          providerId: 'velar-sdk',
-        }),
-      ];
-
-      const result = renderUseSwapState({
-        baseSwapAssets: [btcAsset],
-        targetSwapAssets: [stxAsset],
-        swapQuotes: quotes,
-      });
-
-      act(() => {
-        result.current.actions.setBaseSwapAsset(btcAsset);
-        result.current.actions.setTargetSwapAsset(stxAsset);
-        result.current.actions.setBaseAmount('0.1');
-      });
-
-      await waitFor(() => {
-        expect(result.current.quoteQuery.data).toBeDefined();
-      });
-
-      const enrichedQuotes = result.current.quoteQuery.data?.quotes;
-      assert(enrichedQuotes);
-      expect(enrichedQuotes).toHaveLength(3);
-
-      enrichedQuotes.forEach(enrichedQuote => {
-        expect(enrichedQuote.score).toBe(enrichedQuote.rawSwapQuote.targetAmount);
-      });
-    });
-  });
-
-  describe('network fees', () => {
-    describe('internal fee state management', () => {
-      it('setFeeTier updates feeTier', () => {
-        const result = renderUseSwapState();
-
-        act(() => result.current.actions.setFeeTier('high'));
-        expect(result.current.state.feeTier).toBe('high');
-        expect(result.current.state.customFee).toBeNull();
-      });
-
-      it('setCustomFee updates customFee value', () => {
-        const result = renderUseSwapState();
-
-        expect(result.current.state.customFee).toBeNull();
-
-        act(() => result.current.actions.setCustomFee(15000));
-        expect(result.current.state.customFee).toBe(15000);
-      });
-
-      it('setting tier after custom fee clears customFee', () => {
-        const result = renderUseSwapState();
-
-        act(() => result.current.actions.setCustomFee(20000));
-        expect(result.current.state.customFee).toBe(20000);
-
-        act(() => result.current.actions.setFeeTier('low'));
-        expect(result.current.state.feeTier).toBe('low');
-        expect(result.current.state.customFee).toBeNull();
-
-        act(() => result.current.actions.setCustomFee(10000));
-        expect(result.current.state.customFee).toBe(10000);
-
-        act(() => result.current.actions.setFeeTier('high'));
-        expect(result.current.state.feeTier).toBe('high');
-        expect(result.current.state.customFee).toBeNull();
-      });
-    });
-
-    describe('fee query lifecycle', () => {
-      it('networkFeeQuery is disabled when no quote exists', async () => {
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-          swapQuotes: [],
-        });
-
-        act(() => {
-          result.current.actions.setBaseAmount('0.1');
-        });
-
-        await waitFor(() => {
-          expect(result.current.quoteQuery.status).toBe('success');
-        });
-
-        expect(result.current.quoteQuery.data?.quotes).toEqual([]);
-        expect(result.current.networkFeeQuery.fetchStatus).toBe('idle');
-      });
-
-      it('networkFeeQuery is enabled when valid quote and amount exist', async () => {
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-        });
-
-        act(() => {
-          result.current.actions.setBaseAmount('0.1');
-        });
-
-        await waitFor(() => {
-          expect(result.current.quoteQuery.status).toBe('success');
-          expect(result.current.networkFeeQuery.status).toBe('success');
-        });
-
-        expect(result.current.quoteQuery.data?.selected).toBeDefined();
-        expect(result.current.networkFeeQuery.data).toBeDefined();
-      });
-
-      it('networkFeeQuery refetches when dependencies change', async () => {
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-        });
-
-        act(() => result.current.actions.setBaseAmount('0.1'));
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.status).toBe('success');
-        });
-
-        const firstData = result.current.networkFeeQuery.data;
-
-        act(() => result.current.actions.setBaseAmount('0.2'));
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.data).not.toBe(firstData);
-        });
-
-        const secondData = result.current.networkFeeQuery.data;
-
-        act(() => result.current.actions.setSlippage(0.05));
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.data).not.toBe(secondData);
-        });
-      });
-    });
-
-    describe('fee tier selection', () => {
-      it('default selection is standard tier', async () => {
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-        });
-
-        act(() => result.current.actions.setBaseAmount('0.1'));
-
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.status).toBe('success');
-        });
-
-        const feeData = result.current.networkFeeQuery.data;
-        assert(feeData?.mode === 'tiered');
-        assert(feeData.selected.type === 'tiered');
-        expect(feeData.selected.type).toBe('tiered');
-        expect(feeData.selected.tier).toBe('standard');
-      });
-
-      it.each(['low', 'standard', 'high'] as const)(
-        'selecting %s tier updates state and uses correct fee value',
-        async tier => {
-          const result = renderUseSwapState({
-            baseAsset: defaultBtcAsset,
-            targetAsset: defaultStxAsset,
-          });
-
-          act(() => result.current.actions.setBaseAmount('0.1'));
-
-          await waitFor(() => {
-            expect(result.current.networkFeeQuery.status).toBe('success');
-          });
-
-          act(() => result.current.actions.setFeeTier(tier));
-
-          const feeData = result.current.networkFeeQuery.data;
-          assert(feeData?.mode === 'tiered');
-          assert(feeData.selected.type === 'tiered');
-          expect(feeData.selected.tier).toBe(tier);
-
-          const tierOption = feeData.options.find(opt => opt.tier === tier);
-          assert(tierOption);
-          expect(feeData.calculation.value).toEqual(tierOption.value);
-        }
-      );
-    });
-
-    describe('custom fees', () => {
-      it('setting custom fee changes selection type to custom', async () => {
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-        });
-
-        act(() => result.current.actions.setBaseAmount('0.1'));
-
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.status).toBe('success');
-        });
-
-        act(() => result.current.actions.setCustomFee(25000));
-
-        const feeData = result.current.networkFeeQuery.data;
-        assert(feeData?.mode === 'tiered');
-        assert(feeData.selected.type === 'custom');
-        expect(feeData.selected.type).toBe('custom');
-        expect(feeData.selected.value).toBe(25000);
-      });
-
-      it('custom fee value is reflected in networkFeeQuery.data.value', async () => {
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-        });
-
-        act(() => result.current.actions.setBaseAmount('0.1'));
-
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.status).toBe('success');
-        });
-
-        act(() => result.current.actions.setCustomFee(30000));
-
-        const feeData = result.current.networkFeeQuery.data;
-        assert(feeData?.mode === 'tiered');
-        assert(feeData.selected.type === 'custom');
-
-        expect(feeData.calculation.value.amount.toNumber()).toBe(30000000);
-      });
-
-      it('custom fee takes precedence over tier when both are set', async () => {
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-        });
-
-        act(() => result.current.actions.setBaseAmount('0.1'));
-
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.status).toBe('success');
-        });
-
-        act(() => result.current.actions.setFeeTier('low'));
-
-        let feeData = result.current.networkFeeQuery.data;
-        assert(feeData?.mode === 'tiered');
-        assert(feeData.selected.type === 'tiered');
-        expect(feeData.selected.tier).toBe('low');
-
-        act(() => result.current.actions.setCustomFee(35000));
-
-        feeData = result.current.networkFeeQuery.data;
-        assert(feeData?.mode === 'tiered');
-        assert(feeData.selected.type === 'custom');
-        expect(feeData.selected.value).toBe(35000);
-        expect(feeData.calculation.value.amount.toNumber()).toBe(35000000);
-      });
-    });
-
-    describe('fee and balance interaction', () => {
-      it('fees update when base asset changes', async () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-
-        const stxAsset = createAccountSwapAsset({
-          asset: defaultStxAsset,
-          balance: { crypto: 500_000_000, quote: 5_000_00 },
-        });
-
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-          baseSwapAssets: [btcAsset, stxAsset],
-        });
-
-        act(() => result.current.actions.setBaseAmount('0.1'));
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.status).toBe('success');
-        });
-
-        const firstFeeData = result.current.networkFeeQuery.data;
-
-        act(() => result.current.actions.setBaseSwapAsset(stxAsset));
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.data).not.toBe(firstFeeData);
-        });
-      });
-
-      it('fees update when base amount changes', async () => {
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-        });
-
-        act(() => result.current.actions.setBaseAmount('0.1'));
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.status).toBe('success');
-        });
-
-        const firstFeeData = result.current.networkFeeQuery.data;
-
-        act(() => result.current.actions.setBaseAmount('0.2'));
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.data).not.toBe(firstFeeData);
-        });
-      });
-
-      it('fees remain idle when amount is zero', async () => {
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-        });
-
-        expect(result.current.state.baseAmount).toBe('0');
-
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.fetchStatus).not.toBe('success');
-        });
-      });
-    });
-
-    describe('edge cases', () => {
-      it('changing base asset resets fee tier to standard', async () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-
-        const stxAsset = createAccountSwapAsset({
-          asset: defaultStxAsset,
-          balance: { crypto: 500_000_000, quote: 5_000_00 },
-        });
-
-        const result = renderUseSwapState({
-          baseAsset: defaultBtcAsset,
-          targetAsset: defaultStxAsset,
-          baseSwapAssets: [btcAsset, stxAsset],
-        });
-
-        act(() => result.current.actions.setBaseAmount('0.1'));
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.status).toBe('success');
-        });
-
-        act(() => result.current.actions.setFeeTier('high'));
-        expect(result.current.state.feeTier).toBe('high');
-
-        act(() => result.current.actions.setBaseSwapAsset(stxAsset));
-        expect(result.current.state.feeTier).toBe('standard');
-      });
-
-      it('flipping assets resets fee tier to standard', async () => {
-        const btcAsset = createAccountSwapAsset({
-          asset: defaultBtcAsset,
-          balance: { crypto: 100_000_000, quote: 50_000_00 },
-        });
-
-        const stxAsset = createAccountSwapAsset({
-          asset: defaultStxAsset,
-          balance: { crypto: 500_000_000, quote: 5_000_00 },
-        });
-
-        const result = renderUseSwapState({
-          baseSwapAssets: [btcAsset, stxAsset],
-          targetSwapAssets: [btcAsset, stxAsset],
-        });
-
-        act(() => {
-          result.current.actions.setBaseSwapAsset(btcAsset);
-          result.current.actions.setTargetSwapAsset(stxAsset);
-          result.current.actions.setBaseAmount('0.1');
-        });
-
-        await waitFor(() => {
-          expect(result.current.networkFeeQuery.status).toBe('success');
-        });
-
-        act(() => result.current.actions.setFeeTier('low'));
-        expect(result.current.state.feeTier).toBe('low');
-
-        act(() => result.current.actions.flipAssets());
-        expect(result.current.state.feeTier).toBe('standard');
-      });
-    });
-  });
-});
-
-describe('determineSwapExecutability', () => {
-  it('returns true when all conditions satisfied', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: true },
-      isQuoteFetching: false,
-      quoteBaseAmount: 0.5,
-      currentInput: createMoney(50_000_000, 'BTC', 8),
-    });
-    expect(result).toBe(true);
-  });
-
-  it('returns false when validation invalid', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: false },
-      isQuoteFetching: false,
-      quoteBaseAmount: 0.5,
-      currentInput: createMoney(50_000_000, 'BTC', 8),
-    });
-    expect(result).toBe(false);
-  });
-
-  it('returns false when quote is fetching', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: true },
-      isQuoteFetching: true,
-      quoteBaseAmount: 0.5,
-      currentInput: createMoney(50_000_000, 'BTC', 8),
-    });
-    expect(result).toBe(false);
-  });
-
-  it('returns false when quoteBaseAmount is undefined', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: true },
-      isQuoteFetching: false,
-      quoteBaseAmount: undefined,
-      currentInput: createMoney(50_000_000, 'BTC', 8),
-    });
-    expect(result).toBe(false);
-  });
-
-  it('returns false when currentInput is null', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: true },
-      isQuoteFetching: false,
-      quoteBaseAmount: 0.5,
-      currentInput: null,
-    });
-    expect(result).toBe(false);
-  });
-
-  it('returns true when amounts align for BTC 8 decimals', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: true },
-      isQuoteFetching: false,
-      quoteBaseAmount: 0.12345678,
-      currentInput: createMoney(12_345_678, 'BTC', 8),
-    });
-    expect(result).toBe(true);
-  });
-
-  it('returns true when amounts align for STX 6 decimals', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: true },
-      isQuoteFetching: false,
-      quoteBaseAmount: 100.123456,
-      currentInput: createMoney(100_123_456, 'STX', 6),
-    });
-    expect(result).toBe(true);
-  });
-
-  it('returns false when amounts differ by 1 satoshi', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: true },
-      isQuoteFetching: false,
-      quoteBaseAmount: 0.5,
-      currentInput: createMoney(50_000_001, 'BTC', 8),
-    });
-    expect(result).toBe(false);
-  });
-
-  it('returns false when amounts completely different', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: true },
-      isQuoteFetching: false,
-      quoteBaseAmount: 1.0,
-      currentInput: createMoney(50_000_000, 'BTC', 8),
-    });
-    expect(result).toBe(false);
-  });
-
-  it('handles zero amounts correctly', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: true },
-      isQuoteFetching: false,
-      quoteBaseAmount: 0,
-      currentInput: createMoney(0, 'BTC', 8),
-    });
-    expect(result).toBe(true);
-  });
-
-  it('handles floating point precision for 0.1', () => {
-    const result = determineSwapExecutability({
-      validation: { isValid: true },
-      isQuoteFetching: false,
-      quoteBaseAmount: 0.1,
-      currentInput: createMoney(10_000_000, 'BTC', 8),
-    });
-    expect(result).toBe(true);
-  });
-});
